Fix zero state, small layout paper cuts
