+++ client/web/src/components/ZeroStates/ListPageZeroState.module.scss
@@ -1,0 +1,25 @@
+.zero-state {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    gap: 0.75rem;
+    margin: 1.5rem auto;
+
+    --light-part: var(--gray-03);
+    --middle-part: var(--gray-04);
+    --dark-part: var(--gray-05);
+
+    :global(.theme-dark) & {
+        --light-part: var(--gray-06);
+        --middle-part: var(--gray-07);
+        --dark-part: var(--gray-08);
+    }
+
+    &-text {
+        color: var(--text-muted);
+        display: flex;
+        flex-direction: column;
+        align-items: center;
+        font-size: 1.25rem;
+    }
+}
