+++ client/web/src/enterprise/app/components/index.ts
@@ -1,1 +1,1 @@
-export { AppNoItemsState, NoReposAddedState } from './zero-states/AppZeroStates'
+export { NoReposAddedState } from './zero-states/AppZeroStates'
